GUI: fix 3D texture crash on 64-bit Linux
[JIRA] NA

[JIRA TYPE] Bug-Native-Critical

[Introduced by] NA

[Issue description]
Running the 3D earth demo (CONFIG_REALTEK_BUILD_REAL_EARTH_3D=y) on
64-bit Linux/WSL crashes with SIGSEGV in gui_get_source_color. Same
build works on Windows.

[Issue reproduction]
1. cd win32_sim && menuconfig ../Kconfig.gui  (enable REAL_EARTH_3D)
2. scons
3. ./win32_sim/gui
Backtrace lands in l3_sw_raster.c gui_get_source_color, dereferencing
image_base ~= 0x556F532C (only low 32 bits of the real 0x5555556F532C
data pointer).

[Cause]
Three sites in Lite3D declared texture data addresses as uint32_t while
assigning (uintptr_t)(image->fill_data)/image->data. On 64-bit hosts
this silently truncates the high 32 bits of the pointer, and later
dereference of the truncated address hits an unmapped page.
- l3_tria_raster.c line 736: local `uint32_t image_addr;` plus 4 caller
  function parameters `uint32_t image_addr` in l3_render_with_texture_*
  and l3_fill_texture_*.
- l3_rect_raster.c line 79: local `uint32_t image_base;`.
- l3_sw_raster.c line 225: local `uint32_t line;` in gui_get_rle_pixel
  (same bug shape, affects L3_IMDC compressed textures; not triggered
  by the earth demo but latent on 64-bit).
Windows builds happen to land pointers in the low 4 GB so truncation
still yields a mapped address; Linux ASLR puts them at 0x555555xxxxxx
and truncation lands in an unmapped hole.

[Solution]
Change the affected locals and function parameters from uint32_t to
uintptr_t so the full pointer value is preserved on 64-bit hosts.
Matches gui_raster_params_t.image_base which is already uintptr_t in
l3_sw_raster.h. Field compressed_addr[] stays uint32_t (it holds
per-row byte offsets < 4 GB, correctly added to a uintptr_t base).

[Validation]
- Rebuilt with `scons` in win32_sim (Linux/WSL, gcc, -O0 -g).
- Ran ./win32_sim/gui with the 3D earth demo enabled for 8 seconds
  under gdb; no SIGSEGV, previously crashed within one frame.
- Swept Lite3D for every (uintptr_t)/pointer cast site landing in a
  narrower integer receiver (uint8/16/32_t locals, params, struct
  fields, return values); the three above were the only hits.
- No functional change on 32-bit targets: uintptr_t collapses to
  uint32_t there.

[Need Cherry Pick to] NA

+++ realgui/3rd/Lite3D/widget/l3_rect_raster.c
@@ -76,7 +76,7 @@ void l3_draw_rect_img_to_canvas(l3_draw_rect_img_t *image, l3_canvas_t *dc, l3_r
         if (!l3_draw_img_target_area(image, dc, rect, &x_start, &x_end, &y_start, &y_end)) { return; }
 
         char input_type = head->type;
-        uint32_t image_base;
+        uintptr_t image_base;
         uint8_t *palette_data = NULL;
 
         // Handle I8 format
